+++ sdk/core/azure-core/test/ut/context.cpp
@@ -92,7 +92,7 @@ TEST(Context, ApplicationContext)
   EXPECT_FALSE(appContext.HasKey("1"));
   EXPECT_FALSE(appContext.HasKey(""));
 
-  auto duration = std::chrono::milliseconds(500);
+  auto duration = std::chrono::milliseconds(250);
   EXPECT_FALSE(appContext.IsCancelled());
   std::this_thread::sleep_for(duration);
   EXPECT_FALSE(appContext.IsCancelled());
@@ -100,16 +100,15 @@ TEST(Context, ApplicationContext)
   appContext.Cancel();
   EXPECT_TRUE(appContext.IsCancelled());
 
-  //AppContext2 is the same context as AppContext
+  // AppContext2 is the same context as AppContext
   //  The context should be cancelled
   Context appContext2 = GetApplicationContext();
   EXPECT_TRUE(appContext.IsCancelled());
-
 }
 
 TEST(Context, IsCancelled)
 {
-  auto duration = std::chrono::milliseconds(500);
+  auto duration = std::chrono::milliseconds(250);
   auto deadline = std::chrono::system_clock::now() + duration;
 
   Context context;
@@ -119,9 +118,52 @@ TEST(Context, IsCancelled)
   EXPECT_TRUE(c2.IsCancelled());
 }
 
+TEST(Context, NestedIsCancelled)
+{
+  auto duration = std::chrono::milliseconds(250);
+  auto deadline = std::chrono::system_clock::now() + duration;
+
+  Context context;
+  auto c2 = context.WithValue("Key", "Value");
+  EXPECT_FALSE(c2.IsCancelled());
+  EXPECT_TRUE(c2.HasKey("Key"));
+  EXPECT_FALSE(context.HasKey("Key"));
+
+  auto c3 = context.WithDeadline(deadline);
+  EXPECT_FALSE(context.IsCancelled());
+  EXPECT_FALSE(c2.IsCancelled());
+  EXPECT_FALSE(c3.IsCancelled());
+  std::this_thread::sleep_for(duration);
+
+  EXPECT_FALSE(context.IsCancelled());
+  EXPECT_FALSE(c2.IsCancelled());
+  EXPECT_TRUE(c3.IsCancelled());
+
+  EXPECT_TRUE(c2.HasKey("Key"));
+  EXPECT_FALSE(context.HasKey("Key"));
+  EXPECT_FALSE(c3.HasKey("Key"));
+}
+
+TEST(Context, CancelWithValue)
+{
+  Context context;
+  auto c2 = context.WithValue("Key", "Value");
+  EXPECT_FALSE(context.IsCancelled());
+  EXPECT_FALSE(c2.IsCancelled());
+  EXPECT_TRUE(c2.HasKey("Key"));
+  EXPECT_FALSE(context.HasKey("Key"));
+
+  c2.Cancel();
+  EXPECT_TRUE(c2.IsCancelled());
+  EXPECT_FALSE(context.IsCancelled());
+
+  EXPECT_TRUE(c2.HasKey("Key"));
+  EXPECT_FALSE(context.HasKey("Key"));
+}
+
 TEST(Context, ThrowIfCancelled)
 {
-  auto duration = std::chrono::milliseconds(500);
+  auto duration = std::chrono::milliseconds(250);
   auto deadline = std::chrono::system_clock::now() + duration;
 
   Context context;